requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart==0.0.6
python-docx==1.1.0
PyPDF2==3.0.1
requests==2.31.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv==1.0.0
aiofiles==23.2.1
openpyxl>=3.1.5
pandas>=2.3.3
chromadb>=0.5.0
sentence-transformers>=2.7.0
torch==2.9.1
